requirements.txt
pandas
requests
streamlit
openpyxl
lxml
python-dotenv
beautifulsoup4
watchdog
pyyaml